log_analysis_demo/log-analysis-gemini/src/api/gemini_client.py
import google.generativeai as genai

class GeminiClient:
    def __init__(self, api_key):
        self.api_key = api_key
        # For simulation, we'll mock the API call instead of configuring real API
        # genai.configure(api_key=self.api_key)
        # self.model = genai.GenerativeModel('gemini-1.5-flash')

import google.generativeai as genai
import os

class GeminiClient:
    def __init__(self, api_key):
        self.api_key = api_key
        # For simulation, we'll mock the API call instead of configuring real API
        # genai.configure(api_key=self.api_key)
        # self.model = genai.GenerativeModel('gemini-1.5-flash')

    def fetch_data(self, prompt, log_file_path=None):
        # Mock response for simulation - read from log file
        print(f"Simulating Gemini API call with prompt: {prompt}")
        
        # Read log data from file
        if log_file_path and os.path.exists(log_file_path):
            with open(log_file_path, 'r') as f:
                log_content = f.read()
        else:
            # Fallback to default if file not found
            log_content = """2023-09-04 10:00:00 INFO User login successful
2023-09-04 10:05:00 WARNING High memory usage detected
2023-09-04 10:10:00 ERROR Database connection failed
2023-09-04 10:15:00 INFO Backup completed successfully
2023-09-04 10:20:00 ERROR Invalid user input
2023-09-04 10:25:00 WARNING Disk space low
2023-09-04 10:30:00 INFO System restart initiated"""
        
        # Create a mock response object
        class MockResponse:
            def __init__(self, text):
                self.text = text
        
        mock_response = MockResponse(log_content)
        return self.process_response(mock_response)

    def process_response(self, response):
        # Process the response: extract text and split into log lines
        if response and response.text:
            log_lines = response.text.strip().split('\n')  # Split into lines
            return log_lines  # Return as a list of log entries
        return []