log_analysis_demo/log-analysis-gemini/requirements.txt

requests
python-dotenv
pytest
google-generativeai